statedGraph divided to StatetGraph & ComplexGraph
